refactor:useNewDecorateComponent 제거
* 새로운 꾸미기 컴포넌트 생성 완료 기능 담당하는 useCompleteCreation 를 제거하고, 해당 기능을useNewDecorateComponent 에서 담당하도록 수정
